Fix issue where cursor becomes 'detached' when resizing a widget back and forth [CMP-3]
Issue caused by arbitrary threshold when reducing the size of a widget, replaced this with the column width value and it works more accurately.

There is still a small degree of cursor detachment caused by the browser's inability to track fast movement on a pixel-by-pixel basis, it's unreasonable for us to even attempt to fix this as it'd probably cause performance issues.

The issue where it is no longer possible to increase a widget to the full width of the grid is now so hard to replicate, it's to be considered an edge case (before, it was easy to cause, and needed fixing).