Add ability for attributeParser to accept a ‘tag’ argument
This will include the tag at the start of the output, and allow the parser to modify attributes when working with specific tags.

Example:

```
attributes({'disabled'}, {'tag': (‘a’)})
```

The above example will let us pass the disabled option as normal (to allow us to create the necessary ARIA attributes and classes), but we can test the tag argument to then strip the `disabled` boolean from also being output for `<a>` elements as it's invalid HTML and will throw a W3C validation error.