Restore remove icons for Select2 elements in IE
Issue down to the way IE handles visibility on parents. Seems you can't set a parent to be hidden and a child to be visible.

Fixed by 'hiding' the original X character by setting it to be the same colour as the background and stopping it from becoming visible again through hover.