Add Auto-Adjusting Nav Items List Functionality
This commit helps us detect the window height and calculate how many nav items there can be without the primary nav scrolling. Then adds a “More” nav item and hides all the nav items that cannot fit in. The nav main sliding contains the same list of nav items. Here we calculate how many items are visible in the nav primary and hide the same number of items here. The result should be what can’t be visible in the first to be visible in the second and the opposite.