Fix select2 dropdown positioning
Moves the dropdown elements within the `.contents` element and overrides the inline calculated left position.

Requires relative positioning on the `.controls`, couldn’t spot any issues with this across lexicon but if we wanted to constrain this more we’d have to add a class to the main form__group like we do with checkbox/radio helpers.