Stop select2 elements needing two clicks to open in Firefox
The `overflow: visible` property was preventing the select2 from opening on first click, this was originally added to support the recent primary blue outline on focused elements.

I’ve removed the overflow property and restored the outline the way it should have been done in the beginning, by adding it to the `—open` state of the select2.