Stop default link underline from leaking to naked button styling
Example, `.table [href]` applies the consistent dotted link underline but we don’t particularly want them to leak to naked buttons at this time.

We will need to have a discussion about the a11y affordance of naked buttons, which will likely fail the scruitiny of a WCAG audit, but for now this hotfix is required to maintain existing visuals.